PartialReciprocal: add x86-64-friendly variant
x86-64 doesn't have saturating arithmetic, so LLVM ends up emulating
our saturating add with `add / cmov`, which means it also has to emit
code to fill a register with -1.  However, we know we'll only ever add
1 or 0, so we can compensate for an overflow by subtracting one.

LLVM recognizes the idiom of a conditional subtraction by 1 on
overflow and correctly translates it to a SBB.  On x86-64, this shaves
7 bytes off the generated code, without impact performance.

TESTED=now call both implementations in unit tests.